refactor: inline shell policy stages
